requirements.txt
flask
flask-login
flask-sqlalchemy
werkzeug
groq
python-dotenv
